+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Global authenticator detection for `spring` `SecurityFilterChain` and `configure`
 - `--global` CLI flag for `viz` to enable global authenticator detection
 - `--interprocedural` CLI flag for `viz` to enable interprocedural authenticator detection
+- Apache Shiro authn/authz support for `jax-rs`
+- Apache Shiro authn/authz support for `spring`
 
 ## [0.5.0] - 2023-03-14